+++ SYSLIB/Src/wb_uart.c
@@ -46,8 +46,8 @@ UINT8 uart_rx[RX_ARRAY_NUM] = {0};
 VOID _PutChar_f(UINT8 ucCh);
 UINT32 volatile rx_cnt = 0;
 
-#define sysTxBufReadNextOne()	(((_sys_uUartTxHead+1)==UART_BUFFSIZE)? NULL: _sys_uUartTxHead+1)
-#define sysTxBufWriteNextOne()	(((_sys_uUartTxTail+1)==UART_BUFFSIZE)? NULL: _sys_uUartTxTail+1)
+#define sysTxBufReadNextOne()	(((_sys_uUartTxHead+1)==UART_BUFFSIZE)? (UINT32)NULL: _sys_uUartTxHead+1)
+#define sysTxBufWriteNextOne()	(((_sys_uUartTxTail+1)==UART_BUFFSIZE)? (UINT32)NULL: _sys_uUartTxTail+1)
 #define UART_BUFFSIZE	256
 UINT8 _sys_ucUartTxBuf[UART_BUFFSIZE];
 UINT32 volatile _sys_uUartTxHead, _sys_uUartTxTail;
@@ -326,13 +326,13 @@ INT32 sysInitializeUART(WB_UART_T *uart)
 	// hook UART interrupt service routine
 	if (u32UartPort)
 	{//==1 NORMAL UART
-		_sys_uUartTxHead = _sys_uUartTxTail = NULL;
+		_sys_uUartTxHead = _sys_uUartTxTail = (UINT32)NULL;
 		_sys_pvOldUartVect = sysInstallISR(IRQ_LEVEL_1, IRQ_UART, (PVOID)sysUartISR);
 		sysEnableInterrupt(IRQ_UART);		
 	}
 	else
 	{//==0 High SPEED
-		_sys_uUartTxHead = _sys_uUartTxTail = NULL;
+		_sys_uUartTxHead = _sys_uUartTxTail = (UINT32)NULL;
 		_sys_pvOldUartVect = sysInstallISR(IRQ_LEVEL_1, IRQ_HUART, (PVOID)sysUartISR);
 		sysEnableInterrupt(IRQ_HUART);
 	}
@@ -572,19 +572,19 @@ VOID sysPrintf(PINT8 pcStr,...)
 	WB_UART_T uart;
 	INT8  *argP;
 
-  _sys_bIsUseUARTInt = TRUE;
-	if (!_sys_bIsUARTInitial)
-	{
-    sysUartPort(1);
-		uart.uart_no = WB_UART_1;
+    _sys_bIsUseUARTInt = TRUE;
+    if (!_sys_bIsUARTInitial)
+    {
+        sysUartPort(1);
+        uart.uart_no = WB_UART_1;
 		uart.uiFreq = sysGetExternalClock();
-		uart.uiBaudrate = 115200;		
-		uart.uiDataBits = WB_DATA_BITS_8;
-		uart.uiStopBits = WB_STOP_BITS_1;
-		uart.uiParity = WB_PARITY_NONE;
-		uart.uiRxTriggerLevel = LEVEL_1_BYTE;
-		sysInitializeUART(&uart);		
-   }
+        uart.uiBaudrate = 115200;		
+        uart.uiDataBits = WB_DATA_BITS_8;
+        uart.uiStopBits = WB_STOP_BITS_1;
+        uart.uiParity = WB_PARITY_NONE;
+        uart.uiRxTriggerLevel = LEVEL_1_BYTE;
+        sysInitializeUART(&uart);		
+    }
     
 	vaStart(argP, pcStr);       /* point at the end of the format string */
 	while (*pcStr)
@@ -605,7 +605,7 @@ VOID sysprintf(PINT8 pcStr,...)
 	_sys_bIsUseUARTInt = FALSE;
 	if (!_sys_bIsUARTInitial)
 	{//Default use external clock 12MHz as source clock. 
-    sysUartPort(1);
+		sysUartPort(1);
 		uart.uart_no = WB_UART_1;
 		uart.uiFreq = sysGetExternalClock();
 		uart.uiBaudrate = 115200;
